Refactor question handling in variant generation
